Start setting up for ota development.Undo before merging!
This commit setups up the project to use the at25sf flash as the data
flash. Setup according to the smartfusion2 maker kit, so on the
MSS_SPI_0.

With this the flash test and filesystem tests are both working.

Make sure to setup the flash and fs properly before testing on different setups.