Almost there. Sucessful upload and partial verify.
-Fixed some bugs with the upload logic, and the flash_copy_file.
-Now I have succesfully upload a fw image and verify the checksum, then
copy from the temp slot to the proper fw location.
-Also the verify (cdhArmFw) was tested partially and passed, given I
only upload one image.
-Due to some reliability with the uppload (kiss crc failures) I modify
the terminal to 150 ms between send packets.
-Also the command to reset the fw manager seems to work, and I was able
to have successful upload after a failed one and using the command to
reset the manager.

Next step is to upload golden and update images, just to double check, otherwise
the last step to be verified is the copy from data flash to prog flash,
which must be done on breadboard, or FM. Other steps after would not
have changed.

Also we need to test with different sized fw to check the second path in
the code for writing the final fw packet. (150*num_packets%256 < 106),
where num_packets is ceil(filesize/150).