Tested with maker kit.
Now fully tested as much as possible with the maker kit.
Able to upload and ARM the fw.

-Also tested both pathways for the last packet upload.

So at this point the upload and arm work. Only the copy from raw data flash
to the prog flash needs to be checked othrwise everything after is
unchanged.

Also the filesystem version still builds and appears to work, so this
can be merged in, and the decision made later using the define.