Add saftey check can driver csp queue.
This makes sure the can driver doesn't try and insert into the csp can
rx queue if it is null.

Main use case is if cdh receives any csp message during boot up; after
can initialized but before csp is running. In this case the queue is not
initialized and will hang the software with assert in freertos.