Change fw upload ack seq_num...
-Change the fw upload process so that the cdh will send the expected seq
num, not the last received. This makes it easier for the terminal to
resume after pausing and otherwise should be equivalent.

-Start adding a checksum_file_area() function so we can checksum part of
a file. This is not tested or complete, but should be basically mashup of the current
checksum file and the chucksum program flash area. The function will be
used for locating the bad/missing chunk if the full file checksum fails,
so we can only send the problem chunks vs the whole file.